+++ tests/govtool-frontend/playwright/lib/pages/proposalDiscussionPage.ts
@@ -1,8 +1,8 @@
 import { expect, Locator, Page } from "@playwright/test";
-import { ProposalCreateRequest, ProposedGovAction } from "@types";
+import { ProposalCreateRequest, ProposalType, ProposedGovAction } from "@types";
 import environments from "lib/constants/environments";
 import ProposalDiscussionDetailsPage from "./proposalDiscussionDetailsPage";
-import { PROPOSAL_TYPE_FILTERS } from "@constants/index";
+import { functionWaitedAssert, waitedLoop } from "@helpers/waitedLoop";
 
 export default class ProposalDiscussionPage {
   // Buttons
@@ -32,15 +32,6 @@ export default class ProposalDiscussionPage {
     await this.page.waitForTimeout(2_000);
   }
 
-  async closeUsernamePrompt() {
-    await this.page.waitForTimeout(5_000);
-    await this.page
-      .locator("div")
-      .filter({ hasText: /^Hey, setup your username$/ })
-      .getByRole("button")
-      .click();
-  }
-
   async viewFirstProposal(): Promise<ProposalDiscussionDetailsPage> {
     await this.page
       .locator('[data-testid^="proposal-"][data-testid$="-view-details"]')
@@ -50,7 +41,13 @@ export default class ProposalDiscussionPage {
   }
 
   async getAllProposals() {
-    await this.page.waitForTimeout(4_000); // waits for proposals to render
+    await waitedLoop(async () => {
+      return (
+        (await this.page
+          .locator('[data-testid^="proposal-"][data-testid$="-card"]')
+          .count()) > 0
+      );
+    });
 
     return this.page
       .locator('[data-testid^="proposal-"][data-testid$="-card"]')
@@ -111,7 +108,7 @@ export default class ProposalDiscussionPage {
 
   async clickRadioButtonsByNames(names: string[]) {
     for (const name of names) {
-      const testId = PROPOSAL_TYPE_FILTERS.includes(name)
+      const testId = Object.values(ProposalType).includes(name as ProposalType)
         ? name.toLowerCase()
         : name.toLowerCase().replace(/ /g, "-");
       await this.page.getByTestId(`${testId}-radio`).click();
@@ -130,12 +127,21 @@ export default class ProposalDiscussionPage {
     filters: string[],
     validateFunction: (proposalCard: any, filters: string[]) => Promise<boolean>
   ) {
-    const proposalCards = await this.getAllProposals();
+    await functionWaitedAssert(async () => {
+      const proposalCards = await this.getAllProposals();
 
-    for (const proposalCard of proposalCards) {
-      const hasFilter = await validateFunction(proposalCard, filters);
-      expect(hasFilter).toBe(true);
-    }
+      for (const proposalCard of proposalCards) {
+        const type = await proposalCard
+          .getByTestId("governance-action-type")
+          .textContent();
+        const hasFilter = await validateFunction(proposalCard, filters);
+        if (!hasFilter) {
+          const errorMessage = `A governance action type ${type} does not contain on ${filters}`;
+          throw errorMessage;
+        }
+        expect(hasFilter).toBe(true);
+      }
+    });
   }
 
   async sortAndValidate(