+++ tests/govtool-frontend/playwright/tests/4-proposal-visibility/proposalVisibility.spec.ts
@@ -8,8 +8,10 @@ import {
 } from "@helpers/featureFlag";
 import GovernanceActionDetailsPage from "@pages/governanceActionDetailsPage";
 import GovernanceActionsPage from "@pages/governanceActionsPage";
-import { expect, test } from "@playwright/test";
+import { expect } from "@playwright/test";
+import { test } from "@fixtures/walletExtension";
 import { GovernanceActionType, IProposal } from "@types";
+import { injectLogger } from "@helpers/page";
 
 test.beforeEach(async () => {
   await setAllureEpic("4. Proposal visibility");
@@ -70,15 +72,17 @@ test("4K. Should display correct vote counts on governance details page for disc
   await Promise.all(
     uniqueProposalTypes.map(async (proposalToCheck) => {
       const newPage = await browser.newPage();
+      injectLogger(newPage);
       const govActionDetailsPage = new GovernanceActionDetailsPage(newPage);
       await govActionDetailsPage.goto(
         `${proposalToCheck.txHash}#${proposalToCheck.index}`
       );
 
-      const dRepNotVoted = await govActionDetailsPage.getDRepNotVoted(
-        proposalToCheck,
-        metricsResponsePromise
-      );
+      const dRepTotalAbstainVote =
+        await govActionDetailsPage.getDRepTotalAbstainVoted(
+          proposalToCheck,
+          metricsResponsePromise
+        );
 
       // check dRep votes
       if (await areDRepVoteTotalsDisplayed(proposalToCheck)) {
@@ -86,17 +90,11 @@ test("4K. Should display correct vote counts on governance details page for disc
           `₳ ${correctVoteAdaFormat(proposalToCheck.dRepYesVotes)}`
         );
         await expect(govActionDetailsPage.dRepAbstainVotes).toHaveText(
-          `₳ ${correctVoteAdaFormat(proposalToCheck.dRepAbstainVotes)}`
+          `₳ ${correctVoteAdaFormat(dRepTotalAbstainVote)}`
         );
         await expect(govActionDetailsPage.dRepNoVotes).toHaveText(
           `₳ ${correctVoteAdaFormat(proposalToCheck.dRepNoVotes)}`
         );
-
-        if (dRepNotVoted) {
-          await expect(govActionDetailsPage.dRepNotVoted).toHaveText(
-            `₳ ${correctVoteAdaFormat(dRepNotVoted)}`
-          );
-        }
       }
       // check sPos votes
       if (await areSPOVoteTotalsDisplayed(proposalToCheck)) {