+++ tests/govtool-frontend/playwright/tests/delegation.teardown.ts
@@ -3,7 +3,7 @@ import { adaHolderWallets } from "@constants/staticWallets";
 import { setAllureStory, setAllureEpic } from "@helpers/allure";
 import { skipIfMainnet, skipIfNotHardFork } from "@helpers/cardano";
 import { pollTransaction } from "@helpers/transaction";
-import { test as cleanup } from "@playwright/test";
+import { test as cleanup } from "@fixtures/walletExtension";
 import kuberService from "@services/kuberService";
 
 cleanup.describe.configure({ timeout: environments.txTimeOut });